feature/commande-view : migration on commande, calcul du ttc et ht

+++ models/commande.js
@@ -72,7 +72,7 @@ module.exports = (sequelize, DataTypes) => {
       com_debut_spectacle: DataTypes.DATEONLY,
       com_fin_spectacle: DataTypes.DATEONLY,
       com_remise: DataTypes.FLOAT,
-      com_ht: DataTypes.DECIMAL,
+      com_ht: DataTypes.FLOAT,
       com_tva: DataTypes.DECIMAL,
       com_port: DataTypes.FLOAT,
       com_frais: DataTypes.FLOAT,
